code-dedupe bifs/sort.go
