+++ go.mod
@@ -13,12 +13,11 @@ module github.com/johnkerl/miller
 go 1.15
 
 require (
-	github.com/goccmack/gocc v0.0.0-20210331093148-09606ea4d4d9 // indirect
 	github.com/kballard/go-shellquote v0.0.0-20180428030007-95032a82bc51
-	github.com/lestrrat-go/strftime v1.0.4
-	github.com/mattn/go-isatty v0.0.12
+	github.com/lestrrat-go/strftime v1.0.5
+	github.com/mattn/go-isatty v0.0.14
 	github.com/pkg/profile v1.6.0
-	github.com/stretchr/testify v1.7.0 // indirect
-	golang.org/x/sys v0.0.0-20210326220804-49726bf1d181
+	github.com/stretchr/testify v1.7.0
+	golang.org/x/sys v0.0.0-20210630005230-0f9fa26af87c
 	golang.org/x/term v0.0.0-20201210144234-2321bbc49cbf
 )
