+++ internal/pkg/mlrval/mlrval_sort_test.go
@@ -1,11 +1,9 @@
-package bifs
+package mlrval
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-
-	"github.com/johnkerl/miller/internal/pkg/mlrval"
 )
 
 // ----------------------------------------------------------------
@@ -24,18 +22,14 @@ import (
 
 func TestComparators(t *testing.T) {
 
-	i10 := mlrval.FromInt(10)
-	i2 := mlrval.FromInt(2)
+	i10 := FromInt(10)
+	i2 := FromInt(2)
 
-	bfalse := mlrval.FromBool(false)
-	btrue := mlrval.FromBool(true)
+	bfalse := FromBool(false)
+	btrue := FromBool(true)
 
-	sabc := mlrval.FromString("abc")
-	sdef := mlrval.FromString("def")
-
-	e := *mlrval.ERROR
-
-	a := *mlrval.ABSENT
+	sabc := FromString("abc")
+	sdef := FromString("def")
 
 	//  - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 	// Within-type lexical comparisons
@@ -43,8 +37,8 @@ func TestComparators(t *testing.T) {
 	assert.Equal(t, 0, LexicalAscendingComparator(bfalse, bfalse))
 	assert.Equal(t, -1, LexicalAscendingComparator(bfalse, btrue))
 	assert.Equal(t, -1, LexicalAscendingComparator(sabc, sdef))
-	assert.Equal(t, 0, LexicalAscendingComparator(&e, &e))
-	assert.Equal(t, 0, LexicalAscendingComparator(&a, &a))
+	assert.Equal(t, 0, LexicalAscendingComparator(ERROR, ERROR))
+	assert.Equal(t, 0, LexicalAscendingComparator(ABSENT, ABSENT))
 
 	//  - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 	// Within-type numeric comparisons
@@ -54,30 +48,30 @@ func TestComparators(t *testing.T) {
 
 	assert.Equal(t, 1, NumericAscendingComparator(btrue, bfalse))
 
-	assert.Equal(t, 0, NumericAscendingComparator(&e, &e))
-	assert.Equal(t, 0, NumericAscendingComparator(&a, &a))
+	assert.Equal(t, 0, NumericAscendingComparator(ERROR, ERROR))
+	assert.Equal(t, 0, NumericAscendingComparator(ABSENT, ABSENT))
 
 	//  - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 	// Across-type lexical comparisons
 
 	assert.Equal(t, -1, LexicalAscendingComparator(i10, btrue)) // "10" < "true"
-	assert.Equal(t, -1, LexicalAscendingComparator(i10, sabc))   // "10" < "abc"
-	assert.Equal(t, 1, LexicalAscendingComparator(i10, &e))     // "10" > "(error)"
+	assert.Equal(t, -1, LexicalAscendingComparator(i10, sabc))  // "10" < "abc"
+	assert.Equal(t, 1, LexicalAscendingComparator(i10, ERROR))  // "10" > "(error)"
 
-	assert.Equal(t, 1, LexicalAscendingComparator(bfalse, sabc)) // "false" > "abc"
-	assert.Equal(t, 1, LexicalAscendingComparator(bfalse, &e))   // "false" > "(error)"
+	assert.Equal(t, 1, LexicalAscendingComparator(bfalse, sabc))  // "false" > "abc"
+	assert.Equal(t, 1, LexicalAscendingComparator(bfalse, ERROR)) // "false" > "(error)"
 
 	//  - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 	// Across-type numeric comparisons
 
 	assert.Equal(t, -1, NumericAscendingComparator(i10, btrue))
 	assert.Equal(t, -1, NumericAscendingComparator(i10, sabc))
-	assert.Equal(t, -1, NumericAscendingComparator(i10, &e))
-	assert.Equal(t, -1, NumericAscendingComparator(i10, &a))
+	assert.Equal(t, -1, NumericAscendingComparator(i10, ERROR))
+	assert.Equal(t, -1, NumericAscendingComparator(i10, ABSENT))
 
 	assert.Equal(t, -1, NumericAscendingComparator(bfalse, sabc))
-	assert.Equal(t, -1, NumericAscendingComparator(bfalse, &e))
-	assert.Equal(t, -1, NumericAscendingComparator(bfalse, &a))
+	assert.Equal(t, -1, NumericAscendingComparator(bfalse, ERROR))
+	assert.Equal(t, -1, NumericAscendingComparator(bfalse, ABSENT))
 
-	assert.Equal(t, -1, NumericAscendingComparator(&e, &a))
+	assert.Equal(t, -1, NumericAscendingComparator(ERROR, ABSENT))
 }
