+++ internal/pkg/version/version.go
@@ -1,4 +1,4 @@
 package version
 
 // STRING is the current Miller major/minor/patch version as a single string.
-var STRING string = "6.0.0-dev"
+var STRING string = "6.0.0-rc"
